+++ src/main/java/YogaStudio/service/ClassService.java
@@ -89,10 +89,10 @@ public class ClassService {
         while(sectionsIterator.hasNext())
         {
             SectionEntity section = sectionsIterator.next();
-            List<Integer> enrolledCustomersIds = enrollmentdao.getEnrolledCustomersIds(section.getId());            
-            Iterator<Integer> enrolledCustomersIdsIterator = enrolledCustomersIds.iterator();
+            List<Long> enrolledCustomersIds = enrollmentdao.getEnrolledCustomersIds(section.getId());            
+            Iterator<Long> enrolledCustomersIdsIterator = enrolledCustomersIds.iterator();
             while(enrolledCustomersIdsIterator.hasNext()){
-                Integer id = enrolledCustomersIdsIterator.next();
+                Long id = enrolledCustomersIdsIterator.next();
                 
                 CustomerEntity customerEntity = customerdao.getCustomer(id);
                 
@@ -120,10 +120,10 @@ public class ClassService {
         while(sectionsIterator.hasNext())
         {
             SectionEntity section = sectionsIterator.next();
-            List<Integer> enrolledCustomersIds = enrollmentdao.getEnrolledCustomersIds(section.getId());            
-            Iterator<Integer> enrolledCustomersIdsIterator = enrolledCustomersIds.iterator();
+            List<Long> enrolledCustomersIds = enrollmentdao.getEnrolledCustomersIds(section.getId());            
+            Iterator<Long> enrolledCustomersIdsIterator = enrolledCustomersIds.iterator();
             while(enrolledCustomersIdsIterator.hasNext()){
-                Integer id2 = enrolledCustomersIdsIterator.next();
+                Long id2 = enrolledCustomersIdsIterator.next();
                 
                 CustomerEntity customerEntity = customerdao.getCustomer(id2);
